Widen the Q8_0 GEMV's lanes as part-0 conversions, so C2 runs it at 1.9x of f32
The integer-dot kernel ran at 0.7x of the f32 GEMV on one thread under C2 -- slower
than the f32 matrix it replaces, on the JIT a stock OpenJDK runs a compiled .class
under -- while Graal ran the same code at 1.45x. The cause is in the JDK, not the
arithmetic: convertShape with a non-zero part is slice(origin) then the part-0
conversion, and slice is two rearranges and a blend over an iota shuffle, which C2
compiles as written (~4 cycles each) and Graal folds into the widening instruction.
The kernel did six of them a block.

The block dot now loads the weights as four 64-bit byte vectors and widens each B2S
as part 0, multiplies against the activation quantized straight into a short[] once
per GEMV, and widens a product vector's upper half by swapping the halves with a
constant rearrange and widening part 0 -- the same four lane sums (lane i = columns
j mod 4 = i), so the defun's bits and the pinning tests hold unchanged. 4096x4096,
one thread, two passes: C2 0.72x -> 1.72-1.91x, Graal 1.42-1.57x as before; the
template copy the same within 0.1x. A two-rows-a-pass shape gains 10% under Graal but
overruns C2's NodeCountInliningCutoff and runs boxed, so one row is the shipped size.

Q8GemvBench keeps the 2026-09-05 kernel as its probe row (the reduce-per-block shape's
numbers stand in .todo/672's README); the shapes measured and the one-JVM-per-shape
runner are .todo/artefacts/706-.../. The SIMD guide's "one double multiply-add per
lane" now says single-float, which is what it is.

+++ src/test/java/am/ik/rontolisp/eval/VecSimdQ8KernelsTest.java
@@ -181,13 +181,13 @@ class VecSimdQ8KernelsTest {
 		x[1] = 2.5f;
 		x[2] = -2.5f;
 		x[3] = 0.5f;
-		byte[] xq = new byte[cols];
+		short[] xq = new short[cols];
 		double[] xs = new double[2];
 		VecSimdKernels.quantizeActivationF(x, 0, cols, xq, xs);
-		assertThat(xq[0]).isEqualTo((byte) 127);
-		assertThat(xq[1]).isEqualTo((byte) 2);
-		assertThat(xq[2]).isEqualTo((byte) -2);
-		assertThat(xq[3]).isEqualTo((byte) 0);
+		assertThat(xq[0]).isEqualTo((short) 127);
+		assertThat(xq[1]).isEqualTo((short) 2);
+		assertThat(xq[2]).isEqualTo((short) -2);
+		assertThat(xq[3]).isEqualTo((short) 0);
 		assertThat(xs[0]).isEqualTo(1.0);
 		assertThat(xs[1]).isEqualTo(0.0);
 		byte[] w = blocks(3, cols, 9);
